backend/services/recommendation_engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
numpy==1.26.0
scipy==1.12.0
scikit-learn==1.4.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.26.0
